Fixes TPL mipmap count

+++ BrawlLib/SSBB/ResourceNodes/Graphics/TPLNode.cs
@@ -128,8 +128,8 @@ namespace BrawlLib.SSBB.ResourceNodes
                 tex->_width = (ushort)t.Width;
                 tex->_height = (ushort)t.Height;
                 tex->_LODBias = t._lodBias;
-                tex->_edgeLODEnable = (short)t._enableEdgeLod;
-                tex->_maxLOD = (short)(t.LevelOfDetail - 1);
+                tex->_edgeLODEnable = (byte)t._enableEdgeLod;
+                tex->_maxLOD = (byte)(t.LevelOfDetail - 1);
                 tex->_minLOD = 0;
 
                 texHdrs += TPLTextureHeader.Size;
@@ -310,8 +310,8 @@ namespace BrawlLib.SSBB.ResourceNodes
                         tex->_width = (ushort)Width;
                         tex->_height = (ushort)Height;
                         tex->_LODBias = _lodBias;
-                        tex->_edgeLODEnable = (short)_enableEdgeLod;
-                        tex->_maxLOD = (short)(LevelOfDetail - 1);
+                        tex->_edgeLODEnable = (byte)_enableEdgeLod;
+                        tex->_maxLOD = (byte)(LevelOfDetail - 1);
                         tex->_minLOD = 0;
 
                         VoidPtr data = (VoidPtr)tex + TPLTextureHeader.Size;
